requirements.txt
pyinstaller>=5.1
pywin32>=304
Pillow>=9.0

